requirements_base.txt
otree-core>=1.0.0
Django==1.8.8 # for heroku, needs to be explicitly in requirements file
python-psycopg2
numpy
